+++ changelog.md
@@ -1,6 +1,7 @@
 # 1.0.7
 - Added QOLMOD as incompatible (breaking)
 - Fixed every gamemode not working with the mod
+- Fixed shortcut button position
 
 # 1.0.6
 - Fixed the mod not working with the new update